DESCRIPTION  Use Excel to analyze the sales based on various product categories.   The dataset in file E-Commerce Dashboard dataset.xlsx contains sales data for different product categories. The following are the features in the dataset:  Order ID	Unique Order ID of a product Order Date	Order Placement Date Ship Date	Shipment Date of the placed order Aging	Used to Create Histogram Bin Ship Mode	Shipment mode of placed order Product Category	Product Category Product	Name of the Product Sales	Sales Amount Quantity	The amount or number of a material Discount	A deduction from the usual cost of something Profit	A financial advantage or benefit Shipping Cost	The amount required to ship the placed order Order Priority	Precedence of placed order Customer ID	Unique Customer ID Customer Name	Name of the Customer City	Unique City Name State	Unique State Name Country	Unique Country Name Region	Especially the part of a country Months	The month of placing the order
The following project tasks are required to be performed in Excel:

Use the saved Sample E-Commerce database
Prepare a table of Sales and Profit month-wise in a working sheet
Prepare the sales table region-wise in the working sheet
Create User Control Combo box for Product Category
Create a Column Chart of the month-wise table and region-wise table
Link the table with a combo box
Create a dashboard.
 

Tools required: Microsoft Excel, Data Analysis Add-in

 

Expected Deliverables:  Design a sales dashboard that analyzes the sales based on various product categories. The company wants to add user control for product category so that users can select a category and can see the trend month-wise and product-wise accordingly.